+++ springdoc-openapi-common/src/main/java/org/springdoc/core/RepositoryRestResourceProvider.java
@@ -1,6 +1,7 @@
 package org.springdoc.core;
 
 import java.util.List;
+import java.util.Map;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import org.springdoc.core.fn.RouterOperation;
@@ -18,4 +19,19 @@ public interface RepositoryRestResourceProvider {
 	 * @return the router operations
 	 */
 	List<RouterOperation> getRouterOperations(OpenAPI openAPI);
+
+	/**
+	 * Gets repository rest controller endpoints.
+	 *
+	 * @return the repository rest controller endpoints
+	 */
+	Map<String, Object> getRepositoryRestControllerEndpoints();
+
+	/**
+	 * Gets handler methods.
+	 *
+	 * @return the handler methods
+	 */
+	Map getHandlerMethods();
+
 }