+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.4] - 2021-02-09
+### Added
+- #1053 - Add @RepositoryRestController documentation generation
+### Changed
+- Upgrade versions: swagger-ui to 3.42.0
+### Fixed
+- #1051 - IllegalStateException in org.springdoc.core.SpringDocConfiguration$OpenApiResourceAdvice.handleNoHandlerFound
+- #1047 - swaggerWelcome `WebFluxProperties' that could not be found
+
 ## [1.5.3] - 2021-01-26
 ### Added
 - Added Support for spring.webflux.base-path